requirements.txt
lark>=1.1
pandas>=2.0
tabulate>=0.9
pytest-cov>=4.0
